feat: export some types so they render in docs
Exports the following:
- `CreateAudioResourceOptions` (parameter type in `createAudioResource`)
- `JoinConfig` (parameter type in `createVoiceConnection` and `VoiceConnection#rejoin`)
- `SpeakingMap` (type of `VoiceReceiver#speaking`) and `SpeakingMapEvents`
This is mainly so that these types show in the documentation.

There are some more unexported types referenced in the docs that I haven't exported because they
seem more internal:
- `Awaited`
- `Edge`
- `Networking`

This commit also documents the overloads for `createAudioResource` and `getVoiceConnections`.
Previously, no documentation would be rendered for these two functions as Typedoc only considers the
documentation for each overload.